+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.4.7
+g
+* Fix a bug where npm packages could crash on Node.js if loaded both through
+  `require()` and `import`.
+
 ## 2.4.6
 
 * Properly mark NPM packages as `"type": "module"` when `pkg.jsEsmExports` is